+++ CHANGELOG.md
@@ -33,3 +33,11 @@ Icon now one parameter and custom menu list has button width
 ## 1.1.1+1
 
 Menu list add hide function
+
+## 1.1.2+1
+
+Change screen height base on keyboard size
+
+## 1.1.3
+
+Fix: Change screen height base on keyboard size